Finetune CSS Style
Modified CSS style to make the app more informative & element propotions more in line with each other.
As of now I think I'll let this project be & move on to the next project, but
I'll come back to this project when I've learned more ways to make it even better (or at least I hope I do).